+++ src/pages/admin/Login.vue
@@ -13,7 +13,7 @@
     </q-card-main>
     <q-card-actions>
       <div class="row justify-center btn-row">
-        <q-btn :loading="getIsLoading" color="amber-3" label="Sign In" class="text-brown-6 q-ma-sm col-10" @click="loginCustomer({username,password,type:'password'})">
+        <q-btn :loading="getIsLoading" color="amber-3" label="Sign In" class="text-brown-6 q-ma-sm col-10" @click="loginAdmin({username,password})">
           <q-spinner-pie slot="loading" size="25px" />
         </q-btn>
       </div>
@@ -41,10 +41,10 @@ export default {
     getLoginLogo() {
       return logoData[this.logo]
     },
-    ...mapGetters('customer', ['getIsLoading']),
+    ...mapGetters('admin', ['getIsLoading']),
   },
   methods: {
-    ...mapActions('customer', ['loginCustomer']),
+    ...mapActions('admin', ['loginAdmin']),
     startAnimation() {
       this.vivus = new Vivus(
         'logo',