+++ CHANGELOG.md
@@ -18,12 +18,14 @@ All notable changes to this project will be documented in this file.
 - Change style reply botton and font size desc  [@john](https://github.com/tri-et/TPTEA/tree/feature/menu-detail)
 - Change wrong routing at news detail page (orders->categories)
 - Rename MyLayout to Default
+- Rename auth-token-admin to auth-token
 
 ## 0.0.1 - 2018-09-17
 ### Fixed
 - Missing images due to improper using of assets images
 - Remove dist/ from .gitignore [#reason](in-order to be able to upload to heroku)
 - Change field fullname to name
+- Apply jwt to _get, _post
 ### Changed
 - Remove pull to refresh
 - Use statics assets instead of google drive image download links
@@ -36,3 +38,4 @@ All notable changes to this project will be documented in this file.
 - News & NewsDetail page
 - Categories Listing page
 - Logins fields validation
+- Login admin